add varification and 2 buttons
